OMAP4: Fix Fastboot doesn't work after warm reboot
On PC, USB connection is not detected after steps:
1. USB cable is disconnected.
2. Warm reboot.
3. Enter fastboot mode.
The reason was that after boot Linux kernel set power down the USB PHY.
Workaround: perform check state of the USB PHY and power on it if needed.

Also fastboot doesn't work when PC detected device and usb cabel was
unplugged and then plugged.
Workaround: if USB-cable disconnected just reset faddr.

Change-Id: Id4f6f811a71cfb34ee09e0525189f0e2f59fd193
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>