Fastboot: Fix timing issue where dload_size is reset
There is a scenario where the dload_size value will get
zero'd out after receiving a command.  This causes the device not to be
able to go into DMA download mode.

So to fix this clear the dload_size only if the download is successful or if
we have received a command

Change-Id: I30797734c5711d19aa243e256e2dc2e941c9c8b1
Signed-off-by: Dan Murphy <dmurphy@ti.com>